Reorganize project
To more closely follow conventions for how a gem's files are organized,
I am moving some things under lib/diffux_ci and the DiffuxCI module.
