+++ bin/diffux
@@ -1,9 +1,6 @@
 #!/usr/bin/env ruby
 
-require 'diffux_ci_utils'
-require 'diffux_ci_action'
-require 'diffux_ci_uploader'
-require 'diffux_ci_version'
+require 'diffux_ci'
 require 'fileutils'
 
 help_text = <<-EOS
@@ -24,22 +21,22 @@ case action
 when 'run'
   Thread.abort_on_exception = true
   Thread.new do
-    require 'diffux_ci_runner'
+    require 'diffux_ci/runner'
     exit
   end
-  require 'diffux_ci_server'
+  require 'diffux_ci/server'
 
 when 'debug'
-  system 'open', DiffuxCIUtils.construct_url('/debug')
-  require 'diffux_ci_server'
+  system 'open', DiffuxCI::Utils.construct_url('/debug')
+  require 'diffux_ci/server'
 
 when 'review'
-  system 'open', DiffuxCIUtils.construct_url('/review')
-  require 'diffux_ci_server'
+  system 'open', DiffuxCI::Utils.construct_url('/review')
+  require 'diffux_ci/server'
 
 when 'clean'
-  if File.directory? DiffuxCIUtils.config['snapshots_folder']
-    FileUtils.remove_entry_secure DiffuxCIUtils.config['snapshots_folder']
+  if File.directory? DiffuxCI::Utils.config['snapshots_folder']
+    FileUtils.remove_entry_secure DiffuxCI::Utils.config['snapshots_folder']
   end
 
 when 'approve', 'reject'
@@ -47,11 +44,11 @@ when 'approve', 'reject'
   abort 'Missing example description' unless example_description
   viewport_name = ARGV[2]
   abort 'Missing viewport name' unless viewport_name
-  DiffuxCIAction.new(example_description, viewport_name).send(action)
+  DiffuxCI::Action.new(example_description, viewport_name).send(action)
 
 when 'upload_diffs'
   # `upload_diffs` returns a URL to a static html file
-  puts DiffuxCIUploader.new.upload_diffs
+  puts DiffuxCI::Uploader.new.upload_diffs
 
 when '--version'
   puts "diffux_ci version #{DiffuxCI::VERSION}"
