+++ lib/diffux_ci/public/diffux_ci-runner.js
@@ -101,7 +101,7 @@ window.diffux = {
    *
    * @param {String} exampleDescription
    * @param {Function} doneFunc injected by driver.execute_async_script in
-   *   diffux_ci_runner.rb
+   *   diffux_ci/runner.rb
    */
   renderExample: function(exampleDescription, doneFunc) {
     try {
