+++ lib/diffux_ci/version.rb
