+++ lib/diffux_ci/views/debug.erb
