+++ lib/diffux_ci/views/index.erb
